+++ cmd/psweb/config/lnd.go
@@ -43,6 +43,7 @@ func LoadPS() {
 	port := GetPeerswapLNDSetting("elementsd.rpcport")
 	if port != "" {
 		Config.ElementsPort = port
+		ElementsPortConfigured = true
 	}
 
 	// on the first start without config there will be no elements user and password
@@ -132,7 +133,7 @@ func SavePS() {
 	t += setPeerswapdVariable("lnd.tlscertpath", filepath.Join(Config.LightningDir, "tls.cert"), "", "")
 	t += setPeerswapdVariable("lnd.macaroonpath", filepath.Join(Config.LightningDir, "data", "chain", "bitcoin", Config.Chain, "admin.macaroon"), "", "LND_MACAROONPATH")
 
-	if Config.ElementsPass == "" || Config.ElementsUser == "" {
+	if !Config.LiquidEnabled || Config.ElementsPass == "" || Config.ElementsUser == "" {
 		// disable Liquid so that peerswapd does not fail
 		t += "liquidswaps=false\n"
 		// enable Bitcoin swaps because both cannot be disabled